add motor_mock
