--- exercises/door-polymorphic/leo.traussnigg/lib/motor.h
@@ -1,25 +1,0 @@
-#pragma once
-
-
-class Motor
-{
-public:
-    enum class Direction
-    {
-        IDLE,
-        FORWARD,
-        BACKWARD,
-    };
-
-    Motor(Direction direction);
-    void forward();
-    void backward();
-    void stop();
-
-    // for tests only? (Door uses for a sanity check)
-    Direction get_direction() const;
-
-private:
-    Direction _direction;
-};
-
