+++ exercises/door-polymorphic/leo.traussnigg/tests/CMakeLists.txt
@@ -8,7 +8,7 @@ add_executable(
   door-tests.cpp
 
   # --- this is for you :* ---
-  # motor-polymorphic-tests.cpp
+  motor-polymorphic-tests.cpp
   # lightbarrier-polymorphic-tests.cpp
   # pushbutton-polymorphic-tests.cpp
 )
